Dfr 3601 express pilot dynamic page content (#2208)
* Moved contested solicitor create mid event handler to the same package as other solicitor create handlers

* SOLICITOR_CREATE code in mid event now, so removed from abs handler;
