+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/solicitorcreatecase/SolicitorCreateContestedMidHandler.java
@@ -1,16 +1,20 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.handler;
+package uk.gov.hmcts.reform.finrem.caseorchestration.handler.solicitorcreatecase;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.finrem.caseorchestration.ccd.callback.CallbackType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.controllers.GenericAboutToStartOrSubmitCallbackResponse;
+import uk.gov.hmcts.reform.finrem.caseorchestration.handler.FinremCallbackHandler;
+import uk.gov.hmcts.reform.finrem.caseorchestration.handler.FinremCallbackRequest;
 import uk.gov.hmcts.reform.finrem.caseorchestration.mapper.FinremCaseDetailsMapper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseData;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseDetails;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.FeatureToggleService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.InternationalPostalService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.SelectedCourtService;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.express.ExpressCaseService;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -21,13 +25,19 @@ public class SolicitorCreateContestedMidHandler extends FinremCallbackHandler {
 
     private final InternationalPostalService postalService;
     private final SelectedCourtService selectedCourtService;
+    private final ExpressCaseService expressCaseService;
+    private final FeatureToggleService featureToggleService;
 
     public SolicitorCreateContestedMidHandler(FinremCaseDetailsMapper finremCaseDetailsMapper,
                                               InternationalPostalService postalService,
-                                              SelectedCourtService selectedCourtService) {
+                                              SelectedCourtService selectedCourtService,
+                                              ExpressCaseService expressCaseService,
+                                              FeatureToggleService featureToggleService) {
         super(finremCaseDetailsMapper);
         this.postalService = postalService;
         this.selectedCourtService = selectedCourtService;
+        this.expressCaseService = expressCaseService;
+        this.featureToggleService = featureToggleService;
     }
 
     @Override
@@ -47,6 +57,10 @@ public class SolicitorCreateContestedMidHandler extends FinremCallbackHandler {
         FinremCaseData caseData = caseDetails.getData();
 
         selectedCourtService.setSelectedCourtDetailsIfPresent(caseData);
+
+        if (featureToggleService.isExpressPilotEnabled()) {
+            expressCaseService.setExpressCaseEnrollmentStatus(caseData);
+        }
 
         List<String> errors = new ArrayList<>();
         if (selectedCourtService.royalCourtOrHighCourtChosen(caseData)) {
