+++ src/main/resources/application.properties
@@ -194,14 +194,41 @@ feature.toggle.case_file_view_enabled=${FEATURE_CFV_ENABLED:true}
 feature.toggle.express_pilot_enabled=${EXPRESS_PILOT_ENABLED:false}
 # FRCs in Express Pilot
 finrem.expressCase.frcs=\
-  FR_s_NottinghamList_1,\
-  FR_s_NottinghamList_2,\
-  FR_s_NottinghamList_3,\
-  FR_s_NottinghamList_4,\
-  FR_s_NottinghamList_5,\
-  FR_s_NottinghamList_6,\
-  FR_s_NottinghamList_7,\
-  FR_s_NottinghamList_8
+  FR_birmingham_hc_list_1,\
+  FR_liverpool_hc_list_5,\
+  FR_liverpool_hc_list_2,\
+  FR_liverpool_hc_list_3,\
+  FR_liverpool_hc_list_1,\
+  FR_liverpool_hc_list_4,\
+  FR_cleaveland_hc_list_8,\
+  FR_cleaveland_hc_list_9,\
+  FR_cleaveland_hc_list_2,\
+  FR_cleaveland_hc_list_5,\
+  FR_cleaveland_hc_list_4,\
+  FR_cleaveland_hc_list_1,\
+  FR_cleaveland_hc_list_7,\
+  FR_cleaveland_hc_list_6,\
+  FR_cleaveland_hc_list_3,\
+  FR_nw_yorkshire_hc_list_2,\
+  FR_nw_yorkshire_hc_list_1,\
+  FR_nw_yorkshire_hc_list_3,\
+  FR_nw_yorkshire_hc_list_8,\
+  FR_nw_yorkshire_hc_list_6,\
+  FR_nw_yorkshire_hc_list_7,\
+  FR_nw_yorkshire_hc_list_4,\
+  FR_nw_yorkshire_hc_list_5,\
+  FR_manchester_hc_list_1,\
+  FR_manchester_hc_list_3,\
+  FR_lancashireList_7,\
+  FR_lancashireList_2,\
+  FR_lancashireList_3,\
+  FR_lancashireList_8,\
+  FR_lancashireList_4,\
+  FR_lancashireList_5,\
+  FR_lancashireList_1,\
+  FR_lancashireList_6,\
+  FR_lancashireList_9
+
 # Cron
 cron.batchsize=${CRON_BATCHSIZE:500}
 cron.wait-time-mins=${CRON_WAIT_TIME_MINS:10}
