+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/solicitorcreatecase/SolicitorCreateContestedAboutToSubmitHandlerTest.java
@@ -24,10 +24,8 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.YesOrNo;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.wrapper.ScheduleOneWrapper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.document.DocumentCategory;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.CaseFlagsService;
-import uk.gov.hmcts.reform.finrem.caseorchestration.service.FeatureToggleService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.IdamService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.OnlineFormDocumentService;
-import uk.gov.hmcts.reform.finrem.caseorchestration.service.express.ExpressCaseService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.noc.nocworkflows.UpdateRepresentationWorkflowService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.utils.refuge.RefugeWrapperUtils;
 
@@ -41,7 +39,6 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mockStatic;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -62,10 +59,6 @@ class SolicitorCreateContestedAboutToSubmitHandlerTest {
     CaseFlagsService caseFlagsService;
     @Mock
     IdamService idamService;
-    @Mock
-    FeatureToggleService featureToggleService;
-    @Mock
-    ExpressCaseService expressCaseService;
 
     @Mock
     UpdateRepresentationWorkflowService representationWorkflowService;
@@ -78,8 +71,6 @@ class SolicitorCreateContestedAboutToSubmitHandlerTest {
         handler = new SolicitorCreateContestedAboutToSubmitHandler(
             finremCaseDetailsMapper,
             onlineFormDocumentService,
-            featureToggleService,
-            expressCaseService,
             caseFlagsService,
             idamService,
             representationWorkflowService,
@@ -170,28 +161,6 @@ class SolicitorCreateContestedAboutToSubmitHandlerTest {
             // Check that updateApplicantInRefugeTab is called with our case details instance
             mockedStatic.verify(() -> RefugeWrapperUtils.updateApplicantInRefugeTab(caseDetails), times(1));
         }
-    }
-
-    @Test
-    void testGivenExpressPilotEnabled_ThenExpressCaseServiceCalled() {
-        FinremCallbackRequest callbackRequest = buildFinremCallbackRequest();
-        FinremCaseData caseData = callbackRequest.getCaseDetails().getData();
-        when(featureToggleService.isExpressPilotEnabled()).thenReturn(true);
-
-        handler.handle(callbackRequest, AUTH_TOKEN);
-
-        verify(expressCaseService).setExpressCaseEnrollmentStatus(caseData);
-    }
-
-    @Test
-    void testGivenExpressPilotDisabled_ThenExpressCaseServiceIsNotCalled() {
-        FinremCallbackRequest callbackRequest = buildFinremCallbackRequest();
-        FinremCaseData caseData = callbackRequest.getCaseDetails().getData();
-        when(featureToggleService.isExpressPilotEnabled()).thenReturn(false);
-
-        handler.handle(callbackRequest, AUTH_TOKEN);
-
-        verify(expressCaseService, never()).setExpressCaseEnrollmentStatus(caseData);
     }
 
     private void expectedAdminResponseCaseData(FinremCaseData responseCaseData) {
