+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/solicitorcreatecase/SolicitorCreateContestedMidHandlerTest.java
@@ -7,19 +7,21 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.reform.finrem.caseorchestration.ccd.callback.CallbackType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.handler.FinremCallbackRequest;
-import uk.gov.hmcts.reform.finrem.caseorchestration.handler.SolicitorCreateContestedMidHandler;
 import uk.gov.hmcts.reform.finrem.caseorchestration.mapper.FinremCaseDetailsMapper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseData;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseDetails;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.FeatureToggleService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.InternationalPostalService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.SelectedCourtService;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.express.ExpressCaseService;
+
 
 import java.util.Arrays;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.any;
+import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -39,13 +41,20 @@ class SolicitorCreateContestedMidHandlerTest {
 
     @Mock
     private SelectedCourtService selectedCourtService;
+    @Mock
+    private FeatureToggleService featureToggleService;
+    @Mock
+    private ExpressCaseService expressCaseService;
+
 
     @BeforeEach
     public void init() {
         handler = new SolicitorCreateContestedMidHandler(
                 finremCaseDetailsMapper,
                 postalService,
-                selectedCourtService);
+                selectedCourtService,
+                expressCaseService,
+                featureToggleService);
     }
 
     @Test
@@ -67,6 +76,28 @@ class SolicitorCreateContestedMidHandlerTest {
         handler.handle(finremCallbackRequest, AUTH_TOKEN);
         verify(selectedCourtService, times(1))
                 .setSelectedCourtDetailsIfPresent(finremCallbackRequest.getCaseDetails().getData());
+    }
+
+    @Test
+    void testGivenExpressPilotEnabled_ThenExpressCaseServiceCalled() {
+        FinremCallbackRequest callbackRequest = buildFinremCallbackRequest();
+        FinremCaseData caseData = callbackRequest.getCaseDetails().getData();
+        when(featureToggleService.isExpressPilotEnabled()).thenReturn(true);
+
+        handler.handle(callbackRequest, AUTH_TOKEN);
+
+        verify(expressCaseService).setExpressCaseEnrollmentStatus(caseData);
+    }
+
+    @Test
+    void testGivenExpressPilotDisabled_ThenExpressCaseServiceIsNotCalled() {
+        FinremCallbackRequest callbackRequest = buildFinremCallbackRequest();
+        FinremCaseData caseData = callbackRequest.getCaseDetails().getData();
+        when(featureToggleService.isExpressPilotEnabled()).thenReturn(false);
+
+        handler.handle(callbackRequest, AUTH_TOKEN);
+
+        verify(expressCaseService, never()).setExpressCaseEnrollmentStatus(caseData);
     }
 
     @Test
